Move fs.h to fs_format.h, move inode from file.h to fs.h, move file and fs declarations out of defs.h
