+++ kernel/file.h
@@ -1,3 +1,9 @@
+#pragma once
+
+#include "fs.h"
+#include "pipe.h"
+#include "types.h"
+
 struct file {
   enum { FD_NONE, FD_PIPE, FD_INODE, FD_DEVICE } type;
   int ref; // reference count
@@ -11,23 +17,7 @@ struct file {
 
 #define major(dev)  ((dev) >> 16 & 0xFFFF)
 #define minor(dev)  ((dev) & 0xFFFF)
-#define	mkdev(m,n)  ((uint)((m)<<16| (n)))
-
-// in-memory copy of an inode
-struct inode {
-  uint dev;           // Device number
-  uint inum;          // Inode number
-  int ref;            // Reference count
-  struct sleeplock lock; // protects everything below here
-  int valid;          // inode has been read from disk?
-
-  short type;         // copy of disk inode
-  short major;
-  short minor;
-  short nlink;
-  uint size;
-  uint addrs[NDIRECT+1];
-};
+#define mkdev(m,n)  ((uint)((m)<<16| (n)))
 
 // map major device number to device functions.
 struct devsw {
@@ -38,3 +28,12 @@ struct devsw {
 extern struct devsw devsw[];
 
 #define CONSOLE 1
+
+
+struct file* filealloc(void);
+void         fileclose(struct file*);
+struct file* filedup(struct file*);
+void         fileinit(void);
+int          fileread(struct file*, uint64, int n);
+int          filestat(struct file*, uint64 addr);
+int          filewrite(struct file*, uint64, int n);
