+++ mkfs/mkfs.c
@@ -7,7 +7,7 @@
 
 #define stat xv6_stat  // avoid clash with host struct stat
 #include "kernel/types.h"
-#include "kernel/fs.h"
+#include "kernel/fs_format.h"
 #include "kernel/stat.h"
 #include "kernel/param.h"
